Add support for regional leagues
This allows promotion / relegation of teams into region based leagues.
In some countries lower level leagues are region based, and teams that
are relegated to this level must play in a specific regional league.

To create a regional league, use the <regions> tag in the league
definiton to specify a space separated list of region names that this
league covers.  To specify which region a team belongs to, use the
<region> tag in the team definition.